Fixes the message for MinGW
The messages always prints the "\n" as text instead of treating it as a newline character on Windows with Git Bash and MinGW installed.
In my device (Linux), it works well and if I'm using `-e` flag it would just prints the flag.

So to make it fair, here I changed the newline with implicit one (print blank message with `echo`). I hope this changes fix the issue.