Fixes unable to compile issue
The `javac` command unable to compile, it because the given source files list (`sourceFiles.lst`) has Windows file separator (`\`).
In that case, `javac` treats it as flag then raise an error `invalid flag: "\"`

Please note, the error commonly occurs on Windows. And the recommended way to build the project is by using `Maven`.