Refactoring the `build-docs`
This changes make the `build-docs` rule to be standalone rule,
which means it does not needs and ignores other rules, otherwise raises
an error.

This refactor would make the `FLAGS` options does not conflicting with
other rule that also uses `FLAGS` to add some either additional options or flags (e.g `compile`).
Thus, with refactoring the `build-docs` rule to make it a standalone rule,
it should makes code more easier to maintain on next patch.