The `clean` rule now clean generated docs
Now the ` clean` rule will include the `docs/jmatrix/` directory
during cleaning the working directory and clean it recursively.

And this change also refactored the code, for now on before clean the
specific directory such as `tmp/` or `docs/jmatrix/` it will checks
whether the directory is exist, and cleans the directory only if exist.