Fixed the options issue
This change fixed the issue when both `FLAGS` and `VERBOSE` options
get defined at the same time, the `FLAGS`'s value will get changed by
`VERBOSE`. For example:

    $ make build-docs VERBOSE=all FLAGS="-quiet"

This will not trigger the `javadoc` to use the given FLAGS, instead it
will uses flags that given by `VERBOSE`, which is `-verbose`.