Revamped `codeql` workflow and caching
This change introduces several modifications to the `codeql` workflow:

- The workflow has been reconfigured to exclude Python setup and analysis.
- CodeQL analysis now covers both Ubuntu and Windows platforms to ensure code portability.
- The cache step has been configured to store installed Maven dependencies, improving workflow efficiency.

It's worth noting that the analysis of Python code has been moved to the PyLint workflow.
Additionally, JMatrix's public APIs are primarily built in Java, emphasizing stability, portability, and security.
This shift allows us to focus on enhancing the quality of the Java-based public APIs.