Refactor and enhance the Make's main driver
This refactor change will bring a brand new of `Makefile`. It does not have
any target rules on it, but will send all target rules to `Main.mk` file and
the executions from there. And now the `Makefile` itself supported by the custom flags,
see the usage below.

Usage:
  $ make <TARGET> [CUSTOM_FLAGS]

Alternative usage:
  $ make <TARGET> [FLAGS=<VALUE>]

For example:
  $ make package .lint .include-src .verbose

Equivalent to:
  $ make package LINT=true INCLUDE_SRC=true VERBOSE=true

In the first example, the Make's will trigger and enable LINT, INCLUDE_SRC, and VERBOSE options respectively.
Please use with cautious, the custom flags are experimental options and will lead to unexpected behaviors.