Improve help message and default target
* Updated the help message in "makefile-usage.tx*"
* Users can now customize the behavior of help message

Users can customize the displayed help message by choosing whether to use ANSI color codes or not.
This can be achieved by simply defining the `RAW` or `NOCC` variable with the value `true` on the command line.
For example:

  $ make RAW=true
  $ make NOCC=true

This code will display the help message without ANSI color codes.