Introduce `PropertiesParser` a new parser class
This class provides methods to parse the properties file with ease.
Please note, do not use constructor to create an instance of this class,
use the `load(String)` (static) method instead and use the `getProperties`
method (after created the instance class) to get the instance of
Properties (including all properties data) from this instance.

For example:
    PropertiesParser propsParser = PropertiesParser.load("config.properties");
    Properties configProps = propsParser.getProperties();