Introduce `SetupProperties` subclass in `PropertiesParser`
This class provides access to get the database from the setup properties file.
This class is designed to be synchronized and thread-safe. Its access modifier is "package"
and will only be visible to the package it is stored in.

The setup properties are retrieved from the file statically using the `static` block,
which means that when the program starts it will immediately retrieve the
properties from the file. So this class only provides access to the property data.

Example usage:
    Properties setupProps = SetupProperties.getSetupProperties();