refactor: Add imports checker in Makefile to optimize performance
Added imports checker when importing the Setup.mk, this way the
module will be imported only when the user only when the user
specifies targets, excluding 'help', or when specified nothing.
This approach optimizes performance.

In addition with this commit, refactored the 'help' rule, added
'Makefile' to .PHONY target, and also added license information
on the header message.