feat: Add new method `getEntry` and clarify `get`
Declares and introduces a new method, `getEntry` (including its
static counterpart). Since this method is introduced as the core
method for retrieving elements inside matrix objects, the `get`
method will be an alias of each `getEntry`. We have also added a
static version of the `get` method with 3 parameters (the first
parameter is for the matrix object). In addition, the documentation
of these methods has been added and clarified for API details.

TODO:
Fix the shallow copy of the returned entries array from
`getEntries` method.