deprecation: Deprecate inefficient methods for changing matrix row
Due to inefficiency, the `change` methods for changing row entries in the matrix is being deprecated. Users are advised to transition to using the `insertRow` method for improved performance and functionality.

This change also deprecate the `select` method; as of the deprecation of `change` method, the method are no longer in use and no longer recommended.