refactor: Change the value of `serialVersionUID`
In these huge changes of improving every exceptions in the JMatrix library, we have changed the value of `serialVersionUID` field on every non-deprecated exceptions.

Now the `serialVersionUID` no longer use the calculation provided in changes #60. Instead, now it uses this format:

    430{dd}{MM}{YYYY}{EEE}

Where the {dd} is the date of this project was published to GitHub, {MM} is the month of this project was published to GitHub, {YYYY} is the year of this project was published to GitHub, and the {EEE} is the corresponding error code of the exception (you can check the correct error code for each exceptions in `JMErrorCode` enum).